Move non-component files into lib, and move sidebar into a subdir
