+++ packages/PlanLimitsUI/src/pages/Limits/index.tsx
@@ -4,11 +4,11 @@ import { useDebounce } from 'usehooks-ts';
 import type { ViewState } from 'react-map-gl';
 import { useAtom } from 'jotai';
 import { councilAtom } from '../../lib/loader';
-import { useAppState } from './useAppState';
+import { useAppState } from '../../lib/useAppState';
 import { usePlanLimitsData } from '../../api';
 // import { usePlanLimitsData } from '../../api';
 import Map from './map';
-import Sidebar from './sidebar';
+import Sidebar from './Sidebar';
 import {
   type loader,
   viewLocationUrlPath,
